Extract pure path/artifact logic to downloader-paths.ts
Move vscode-independent functions (getArtifactDownloadName, getClojureLspPath,
getVersionFilePath, readVersionFile) to their own module so tests import the
real implementation instead of duplicating it.

Part of #3211
