+++ src/lsp/client/downloader.ts
@@ -5,10 +5,18 @@ import * as config from '../../config';
 import * as path from 'node:path';
 import * as vscode from 'vscode';
 import * as fs from 'node:fs';
+import {
+  getArtifactDownloadName,
+  getClojureLspPath,
+  getVersionFilePath,
+  readVersionFile,
+} from './downloader-paths';
+
+export { getArtifactDownloadName, getClojureLspPath, getVersionFilePath, readVersionFile };
+
 const DOWNLOAD_TIMEOUT_MS = 120_000;
 const LOCK_STALE_MS = 2 * 60 * 1000; // 2 minutes
 
-const versionFileName = 'clojure-lsp-version';
 const lockFileName = '.downloading';
 
 export function getClojureLspStorageDir(context: vscode.ExtensionContext): string {
@@ -45,52 +53,6 @@ async function acquireDownloadLock(storageDir: string): Promise<boolean> {
 async function releaseDownloadLock(storageDir: string): Promise<void> {
   const lockPath = path.join(storageDir, lockFileName);
   await fs.promises.unlink(lockPath).catch(() => undefined);
-}
-
-const artifacts = {
-  darwin: {
-    x64: 'clojure-lsp-native-macos-amd64.zip',
-    arm64: 'clojure-lsp-native-macos-aarch64.zip',
-  },
-  linux: {
-    x64: 'clojure-lsp-native-static-linux-amd64.zip',
-    arm64: 'clojure-lsp-native-linux-aarch64.zip',
-  },
-  win32: {
-    x64: 'clojure-lsp-native-windows-amd64.zip',
-  },
-};
-
-export function getArtifactDownloadName(
-  platform: string = process.platform,
-  arch: string = process.arch
-): string {
-  return artifacts[platform]?.[arch] ?? 'clojure-lsp-standalone.jar';
-}
-
-export function getClojureLspPath(
-  baseDir: string,
-  platform: string = process.platform,
-  arch: string = process.arch
-): string {
-  let name = getArtifactDownloadName(platform, arch);
-  if (path.extname(name).toLowerCase() !== '.jar') {
-    name = platform === 'win32' ? 'clojure-lsp.exe' : 'clojure-lsp';
-  }
-  return path.join(baseDir, name);
-}
-
-export function getVersionFilePath(baseDir: string): string {
-  return path.join(baseDir, versionFileName);
-}
-
-export async function readVersionFile(baseDir: string) {
-  const filePath = getVersionFilePath(baseDir);
-  try {
-    return await fs.promises.readFile(filePath, 'utf8');
-  } catch (e) {
-    console.error('Could not read clojure-lsp version file.', e.message);
-  }
 }
 
 function downloadArtifact(url: string, filePath: string): Promise<void> {
